rejoins tests
